Lab2: Challenges, Bug fixes

+++ lab/kern/pmap.h
@@ -15,7 +15,7 @@ extern struct PageInfo *pages;
 extern size_t npages;
 
 extern pde_t *kern_pgdir;
-extern uint32_t pse_support;
+extern uint8_t pse_support;
 
 /* This macro takes a kernel virtual address -- an address that points above
  * KERNBASE, where the machine's maximum 256MB of physical memory is mapped --
